Fixes key error and all uncaught exceptions
It was discovered that with large time steps, the
transitions are not necessarily atomic. So for example
a simplex and edge can be removed and yet not have
two boundary cycles merging into one. Or worse, in some
cases boundary cycles are altered with no edges or simpleces
being added or removed.

The solution was to check cases based on the number of added
and removed boundary cycles as well instead of just the
difference in the number of cycles.

The fix appears to catch all previously unhandled errors.